(chore) Moves option mockers into mock directory
