+++ containers/mosaic/src/mock/options.ts
@@ -1,4 +1,5 @@
 import type { MosaicLabelValue } from "@root/types";
+import countriesWithStates from "./trimmedCountriesStates.json";
 
 export const optionsLibrary: MosaicLabelValue[] = [
 	{
@@ -77,4 +78,23 @@ export const getOptions: () => Promise<MosaicLabelValue[]> = async () => {
 			resolve(optionsLibrary);
 		}, 3000);
 	});
+};
+
+export const getOptionsCountries = async (): Promise<MosaicLabelValue[]> => {
+	return Promise.resolve(countriesWithStates.map((country) => ({
+		label: country.name,
+		value: country.iso2,
+	})));
+};
+
+export const getOptionsStates = async (countryValue: string | undefined): Promise<MosaicLabelValue[]> => {
+	const selectedCountry = countriesWithStates?.find(
+		(c) => c.iso2 === countryValue,
+	);
+
+	if (selectedCountry) {
+		return Promise.resolve(selectedCountry.states.map((state) => ({ label: state.name, value: state.code })));
+	}
+
+	return Promise.resolve([]);
 };
