+++ containers/mosaic/src/mock/trimmedCountriesStates.json
